chore: Extract test utils to separate file
This makes it possible to reuse the utils in other test files.
